requirements.txt
fastapi==0.110.3
uvicorn==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.8
tensorflow==2.17.0
streamlit==1.37.0
crewai==0.51.1
crewai-tools==0.8.3
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.12
openai==1.37.1
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.1
seaborn==0.13.2
scikit-learn==1.5.1
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.2
beautifulsoup4==4.12.3
passlib==1.7.4
python-jose==3.3.0
plotly==5.23.0
wordcloud==1.9.3